Revert "code: adapters moved inside detox object"
This reverts commit b3fe368c834b1b19c877770021bfcee6f5c9ce47.
