+++ detox/local-cli/templates/jest.js
@@ -5,19 +5,22 @@ const runnerConfig = `{
 
 const initjsContent = `const detox = require('detox');
 const config = require('../package.json').detox;
+const adapter = require('detox/runners/jest/adapter');
 
 jest.setTimeout(120000);
+jasmine.getEnv().addReporter(adapter);
 
 beforeAll(async () => {
   await detox.init(config);
 });
 
 beforeEach(async () => {
-  await adapter.beforeEach.jest();
+  await adapter.beforeEach();
 });
 
 afterAll(async () => {
-  await detox.cleanup.jest();
+  await adapter.afterAll();
+  await detox.cleanup();
 });`;
 
 exports.initjs = initjsContent;
