+++ detox/package.json
@@ -22,7 +22,7 @@
   "license": "MIT",
   "scripts": {
     "build": "node scripts/build.js",
-    "lint": "eslint src scripts",
+    "lint": "eslint runners src scripts",
     "unit": "jest --coverage --verbose",
     "pretest": "npm run lint",
     "test": "npm run unit",
@@ -69,7 +69,6 @@
       "src/artifacts/video",
       "src/artifacts/templates/plugin",
       "src/artifacts/ArtifactsManager.js",
-      "src/runners",
       ".*Driver.js",
       "EmulatorTelnet.js",
       "Emulator.js",
