+++ detox/runners/jest/adapter.js
@@ -1,4 +1,4 @@
-const detox = require('../../index');
+const detox = require('../../src/index');
 const DetoxJestAdapter = require('./DetoxJestAdapter');
 
 module.exports = new DetoxJestAdapter(detox);
