+++ detox/runners/mocha/adapter.js
@@ -1,4 +1,4 @@
-const detox = require('../../index');
+const detox = require('../../src/index');
 const DetoxMochaAdapter = require('./DetoxMochaAdapter');
 
 module.exports = new DetoxMochaAdapter(detox);
