+++ detox/src/index.js
@@ -5,8 +5,6 @@ const platform = require('./platform');
 const exportWrapper = require('./exportWrapper');
 const argparse = require('./utils/argparse');
 const configuration = require('./configuration');
-const mochaAdapter = require('./runners/mocha/adapter');
-const jestAdapter = require('./runners/jest/adapter').tryToRegisterAsJasmineReporter();
 
 let detox;
 
@@ -61,28 +59,17 @@ async function beforeEach(testSummary) {
   }
 }
 
-beforeEach.mocha = mochaAdapter.beforeEach.bind(mochaAdapter);
-beforeEach.jest = jestAdapter.beforeEach.bind(jestAdapter);
-
 async function afterEach(testSummary) {
   if (detox) {
     await detox.afterEach(testSummary);
   }
 }
 
-afterEach.mocha = mochaAdapter.afterEach.bind(mochaAdapter);
-
 async function cleanup() {
     if (detox) {
         await detox.cleanup();
     }
 }
-
-/* istanbul ignore next */
-cleanup.jest = async () => {
-  await jestAdapter.afterAll();
-  await cleanup();
-};
 
 /* istanbul ignore next */
 const _terminate = _.once(() => detox && detox.terminate());
