+++ examples/demo-react-native/e2e/init.js
@@ -1,16 +1,17 @@
 const detox = require('detox');
 const config = require('../package.json').detox;
+const adapter = require('detox/runners/mocha/adapter');
 
 before(async () => {
   await detox.init(config);
 });
 
 beforeEach(async function () {
-  await detox.beforeEach.mocha(this);
+  await adapter.beforeEach(this);
 });
 
 afterEach(async function () {
-  await detox.afterEach.mocha(this);
+  await adapter.afterEach(this);
 });
 
 after(async () => {
